+++ CHANGELOG.md
@@ -6,6 +6,12 @@ All notable changes to this package will be documented in this file.
 
 ---
 
+## [4.12.13-alpha.1] - 2025-05-30
+
+### 🕧 Experimenting Plugin System for Chronos
+
+- Introduced plugin injection in `Chronos` class. Started with `season` method. Will make convert more methods if this is successful after publishing.
+
 ## [4.12.12] - 2025-05-30
 
 ### 🕧 Updates in Chronos